fix: edit the PID file

+++ tests/pid & ppid/pid.c
@@ -1,16 +1,19 @@
 #include <stdio.h>
 #include <unistd.h>
+#include <sys/types.h>
 
 /**
- * main - PID
+ * main - Prints the PID of the current process
  *
- * Return: Always 0.
+ * Return: Always 0 if success
  */
+
 int main(void)
 {
     pid_t my_pid;
 
     my_pid = getpid();
-    printf("%u\n", my_pid);
+    printf("%d\n", (int)my_pid);
+
     return (0);
 }
